Improve UX flow by allowing users to select a driver directly to start a ride
