+++ src/components/RideServices.jsx
@@ -1,6 +1,5 @@
 import {useState} from "react";
-import FindDriverForm from "./forms/FindDriverForm";
-import StartRideForm from "./forms/StartRideForm.jsx";
+import FindRideForm from "./forms/FindRideForm.jsx";
 import StopRideForm from "./forms/StopRideForm.jsx";
 import BillForm from "./forms/BillForm.jsx";
 import RateDriverForm from "./forms/RateDriverForm.jsx";
@@ -10,15 +9,12 @@ import RechargeForm from "./forms/RechargeForm.jsx";
 const RideServices = () => {
     const [selectedOption, setSelectedOption] = useState(null);
 
-    const options = ["Find Driver", "Start Ride", "Stop Ride", "Generate Bill", "Pay Bill", "Rate Driver", "Recharge Wallet"];
+    const options = ["Find Your Ride", "Stop Ride", "Generate Bill", "Pay Bill", "Rate Driver", "Recharge Wallet"];
 
     const renderComponent = () => {
         switch (selectedOption) {
-            case "Find Driver":
-                return <FindDriverForm />;
-
-            case "Start Ride":
-                return <StartRideForm />;
+            case "Find Your Ride":
+                return <FindRideForm />;
 
             case "Stop Ride":
                 return <StopRideForm />;
