--- src/components/forms/FindDriverForm.jsx
@@ -1,75 +1,0 @@
-import {useState} from "react";
-import axios from "axios";
-import toast from "react-hot-toast";
-
-const FindDriverForm = () => {
-    const [matchedDrivers, setMatchedDrivers] = useState(null);
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-
-        if (localStorage.getItem("RiderID") === null) {
-            toast.error("Please register as a rider first!");
-            return;
-        }
-
-        const request = axios.get(import.meta.env.VITE_BASE_URL + "/ride/rider/match", {
-            params: {
-                riderID: localStorage.getItem("RiderID")
-            }
-        });
-
-        toast.promise(request, {
-            loading: "Searching for potential drivers...",
-            success: "Search Complete!",
-            error: (err) => err.response.data.error
-        });
-
-        request.then(response => {
-            setMatchedDrivers(response.data?.matchedDrivers ?? null);
-        }).catch(error => console.error(error.response));
-    };
-
-
-    return (
-        <div className={"h-[28rem] flex flex-col justify-center items-center text-xl font-bold"}>
-            <form
-                onSubmit={handleSubmit}
-                className="flex flex-col w-2/3 justify-center items-center"
-            >
-                <h2 className={"my-8 justify-center items-center"}>Ready to Find Nearby Drivers?</h2>
-
-                <div className="flex justify-center mt-2">
-                    <button
-                        type="submit"
-                        className="text-lg bg-black text-white px-4 py-2 rounded-lg hover:bg-gray-800"
-                    >
-                        Search for Drivers
-                    </button>
-                </div>
-            </form>
-
-            <div className={"my-8"}>
-                {matchedDrivers !== null ? (
-                    matchedDrivers.length > 0 ? (
-                        <div className={"flex flex-col items-center"}>
-                            <h1 className={"mb-4"}>Your Matched Drivers (Closest to Furthest)</h1>
-                            {matchedDrivers.map((driver, index) => (
-                                <p
-                                    key={index}
-                                    className={"font-medium"}
-                                >
-                                    {index+1}. DRIVER_D{driver}
-                                </p>
-                            ))}
-                        </div>
-                    ) : (
-                        <p>No Drivers Available.</p>
-                    )
-                ) : null}
-            </div>
-        </div>
-    );
-};
-
-export default FindDriverForm;
