--- src/components/forms/StartRideForm.jsx
@@ -1,72 +1,0 @@
-import {useState} from "react";
-import axios from "axios";
-import toast from "react-hot-toast";
-
-const StartRideForm = () => {
-    const [driver, setDriver] = useState("");
-    const [rideStatus, setRideStatus] = useState(null);
-
-    const handleDriverChange = (e) => {
-        const value = e.target.value.replace(/\D/g, "");
-        setDriver(value);
-    };
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-
-        axios.post(import.meta.env.VITE_BASE_URL + "/ride/start", null, {
-            params: {
-                N: driver,
-                riderID: localStorage.getItem("RiderID")
-            }
-        }).then(response => {
-            localStorage.setItem("RideID", response.data.rideID);
-            localStorage.setItem("DriverID", response.data.driverID);
-            setRideStatus(response.data);
-
-            toast.success("Ride Started Successfully!");
-        }).catch(error => {
-            toast.error(error.response.data.error);
-        });
-    };
-
-    return (
-        <div className={"h-[30rem] flex flex-col justify-center items-center text-xl font-bold"}>
-            {!rideStatus ? (
-                <form
-                    onSubmit={handleSubmit}
-                    className="grid grid-cols-4 gap-4 w-2/3 justify-center items-center"
-                >
-                    <label className="col-span-2 font-medium text-right">Choose a driver from your matches:</label>
-                    <input
-                        type="text"
-                        value={driver}
-                        onChange={handleDriverChange}
-                        className="col-span-2 px-2 py-1 rounded-lg border-black/80 border-2"
-                    />
-
-                    <div className="col-span-4 flex justify-center mt-2">
-                        <button
-                            type="submit"
-                            className="bg-black text-white px-4 py-2 rounded-lg hover:bg-gray-800"
-                        >
-                            Start My Ride
-                        </button>
-                    </div>
-                </form>
-            ) : (
-                <div className="mt-8 mb-4 bg-gray-200 rounded-xl flex flex-col items-center">
-                    <h1 className="my-4">Your Ride Details</h1>
-                    <div className="grid grid-cols-2 gap-4 font-medium">
-                        <div className="p-4 text-center">Ride ID: {rideStatus.rideID}</div>
-                        <div className="p-4 text-center">Rider ID: {rideStatus.riderID}</div>
-                        <div className="p-4 text-center">Driver ID: {rideStatus.driverID}</div>
-                        <div className="p-4 text-center">Ride Status: {rideStatus.status}</div>
-                    </div>
-                </div>
-            )}
-        </div>
-    );
-};
-
-export default StartRideForm;
